+++ src/Http/Handlers/ControlMenuHandler.php
@@ -1,4 +1,4 @@
-<?php namespace Orchestra\Control;
+<?php namespace Orchestra\Control\Http\Handlers;
 
 use Orchestra\Foundation\Support\MenuHandler;
 use Orchestra\Contracts\Authorization\Authorization;